test(tui): TestEffectiveProjectId を test_state.py へ移動
テスト対象は TuiState.effective_project_id() /
effective_wiki_project_id() (state.py) であり、モジュール単位で
テストファイルを分ける慣習に合わせて test_state.py へ移す
