+++ tests/unit/tui/test_state.py
@@ -1,3 +1,4 @@
+from redi import config
 from redi.i18n import messages
 from redi.tui.state import IssueFilter, TimeEntryFilter, TuiResult, TuiState
 
@@ -73,6 +74,42 @@ class TestTimeEntryFilter:
         f = TimeEntryFilter(user_id="42", user_label="Alice")
         assert f.is_active() is True
         assert f.short_label() == "user=Alice"
+
+
+class TestEffectiveProjectId:
+    """effective_project_id() は override > config の優先順位で解決する"""
+
+    def test_override_wins(self, monkeypatch):
+        monkeypatch.setattr(config, "default_project_id", "5")
+        monkeypatch.setattr(config, "wiki_project_id", "7")
+        state = TuiState(project_id="2")
+
+        assert state.effective_project_id() == "2"
+        # 明示切替は wiki_project_id より優先する
+        assert state.effective_wiki_project_id() == "2"
+
+    def test_falls_back_to_config(self, monkeypatch):
+        monkeypatch.setattr(config, "default_project_id", "5")
+        monkeypatch.setattr(config, "wiki_project_id", "7")
+        state = TuiState()
+
+        assert state.effective_project_id() == "5"
+        assert state.effective_wiki_project_id() == "7"
+
+    def test_wiki_falls_back_to_default_project(self, monkeypatch):
+        monkeypatch.setattr(config, "default_project_id", "5")
+        monkeypatch.setattr(config, "wiki_project_id", None)
+        state = TuiState()
+
+        assert state.effective_wiki_project_id() == "5"
+
+    def test_none_when_nothing_is_set(self, monkeypatch):
+        monkeypatch.setattr(config, "default_project_id", None)
+        monkeypatch.setattr(config, "wiki_project_id", None)
+        state = TuiState()
+
+        assert state.effective_project_id() is None
+        assert state.effective_wiki_project_id() is None
 
 
 class TestCarryOver:
